+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
-
-_0.3.4 "....." (xx.11.2016)_
+_0.3.5 "Maiao" (xx.12.2016)_
 * ...
+
+
+### 0.3.4 "Fitii" (09.11.2016)
+* Added: trigger `AFTER DELETE`
+* Fix: `TRUNCATE TABLE` now works for local storage DB
+* Fix: `JOIN` a sub select
+* Removed: The `HELP` command (to save space) 
 
 
 ### 0.3.3 "Makemo" (13.10.2016)
@@ -11,6 +17,8 @@ _0.3.4 "....." (xx.11.2016)_
 * Add: Load CSV data from a string
 * Add: Warn when server side uses browser build of lib
 * Update: typescript definition for native import 
+* Update: filesaver.js updated to 1.3.2
+
 
 ### 0.3.2 "Maumu" (05.09.2016)
 * Added: Postgres arrays like array[] and text[]